requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
streamlit==1.32.0
moviepy==1.0.3
openai==1.82.0
yt-dlp==2025.5.22
numpy<2,>=1.19.3
python-multipart==0.0.20
imageio==2.37.0
imageio_ffmpeg==0.6.0
openai-whisper==20240930
